Preserve Daum-Huang particle weights
