--- .github/workflows/agent-preserve-daum-huang-weights.yml
@@ -1,130 +1,0 @@
-name: Apply Daum-Huang weight preservation fix
-
-on:
-  push:
-    branches:
-      - agent/preserve-daum-huang-weights-20260715
-  pull_request:
-    branches:
-      - main
-    types: [opened, synchronize, reopened, ready_for_review]
-
-permissions:
-  contents: write
-
-jobs:
-  apply-fix:
-    if: github.actor != 'github-actions[bot]'
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v4
-        with:
-          fetch-depth: 0
-
-      - name: Patch source and add regression coverage
-        shell: bash
-        run: |
-          python - <<'PY'
-          from pathlib import Path
-
-          source_path = Path("src/pyrecest/filters/daum_huang_particle_filter.py")
-          source = source_path.read_text(encoding="utf-8")
-          old_import = "from pyrecest.backend import asarray, eye, ones, to_numpy"
-          new_import = "from pyrecest.backend import asarray, eye, to_numpy"
-          if source.count(old_import) != 1:
-              raise SystemExit("unexpected Daum-Huang backend import")
-          source = source.replace(old_import, new_import, 1)
-
-          old_state_update = """        n_particles = self.filter_state.w.shape[0]\n        self._filter_state = LinearDiracDistribution(\n            result, ones(n_particles) / n_particles\n        )"""
-          new_state_update = """        self._filter_state = LinearDiracDistribution(\n            result, self.filter_state.w\n        )"""
-          if source.count(old_state_update) != 1:
-              raise SystemExit("unexpected Daum-Huang state update")
-          source = source.replace(old_state_update, new_state_update, 1)
-          source_path.write_text(source, encoding="utf-8")
-
-          test_path = Path("tests/filters/test_daum_huang_particle_filter.py")
-          tests = test_path.read_text(encoding="utf-8")
-          marker = """    def test_ledh_records_particlewise_linearization_points(self):\n"""
-          if tests.count(marker) != 1:
-              raise SystemExit("unexpected Daum-Huang test layout")
-          regression = '''    def test_edh_filter_preserves_nonuniform_particle_weights(self):
-        particles = array(
-            [
-                [-2.0, -1.0],
-                [-1.0, 1.0],
-                [0.0, 0.0],
-                [1.0, -1.0],
-                [2.0, 1.0],
-            ]
-        )
-        prior = LinearDiracDistribution(
-            particles, array([0.05, 0.10, 0.15, 0.25, 0.45])
-        )
-        expected_weights = to_numpy(prior.w).copy()
-        prior_mean = prior.mean()
-        prior_covariance = prior.covariance()
-        measurement_matrix = array([[1.0, -0.5]])
-        measurement = array([0.25])
-        measurement_noise = array([[0.75]])
-        expected_mean, expected_covariance = gaussian_bridge_moments(
-            prior_mean,
-            prior_covariance,
-            measurement_matrix,
-            measurement,
-            measurement_noise,
-            1.0,
-            jitter=0.0,
-        )
-        filt = EDHParticleFlowFilter(
-            n_particles=particles.shape[0],
-            dim=particles.shape[1],
-            n_steps=4,
-            jitter=0.0,
-        )
-        filt.filter_state = prior
-
-        filt.update_linear(measurement, measurement_matrix, measurement_noise)
-
-        npt.assert_allclose(to_numpy(filt.filter_state.w), expected_weights)
-        npt.assert_allclose(
-            to_numpy(filt.filter_state.mean()), to_numpy(expected_mean), atol=1e-10
-        )
-        npt.assert_allclose(
-            to_numpy(filt.filter_state.covariance()),
-            to_numpy(expected_covariance),
-            atol=1e-10,
-        )
-
-'''
-          tests = tests.replace(marker, regression + marker, 1)
-          test_path.write_text(tests, encoding="utf-8")
-
-          Path(".github/workflows/agent-preserve-daum-huang-weights.yml").unlink()
-          PY
-
-      - name: Validate patch structure and syntax
-        shell: bash
-        run: |
-          python -m py_compile \
-            src/pyrecest/filters/daum_huang_particle_filter.py \
-            tests/filters/test_daum_huang_particle_filter.py
-          python - <<'PY'
-          from pathlib import Path
-
-          source = Path("src/pyrecest/filters/daum_huang_particle_filter.py").read_text()
-          tests = Path("tests/filters/test_daum_huang_particle_filter.py").read_text()
-          assert "result, self.filter_state.w" in source
-          assert "ones(n_particles) / n_particles" not in source
-          assert "test_edh_filter_preserves_nonuniform_particle_weights" in tests
-          PY
-
-      - name: Commit fix to pull-request branch
-        shell: bash
-        env:
-          TARGET_BRANCH: ${{ github.head_ref || github.ref_name }}
-        run: |
-          git config user.name "github-actions[bot]"
-          git config user.email "41898282+github-actions[bot]@users.noreply.github.com"
-          git add -A
-          git commit -m "Preserve Daum-Huang particle weights"
-          git push origin "HEAD:${TARGET_BRANCH}"
